+++ src/Credfeto.Aur.Mirror.Git/Exceptions/GitException.cs
@@ -1,14 +1,6 @@
-using System;
+using System.ComponentModel;
 
 namespace Credfeto.Aur.Mirror.Git.Exceptions;
 
-public sealed class GitException : Exception
-{
-    public GitException() { }
-
-    public GitException(string? message)
-        : base(message) { }
-
-    public GitException(string? message, Exception? innerException)
-        : base(message: message, innerException: innerException) { }
-}
+[Description("Git operation failed.")]
+public sealed partial class GitException : global::System.Exception;